requirements.txt
fastapi
uvicorn
requests
PyPDF2
PyMuPDF
pinecone
azure-ai-inference
python-dotenv
openai
google-generativeai